ref: extract command operations into commands.js
