+++ index.js
@@ -1,8 +1,7 @@
 import { createInterface } from 'node:readline/promises';
 import { homedir } from 'node:os';
-import { join, sep } from 'node:path';
-import { readdir, stat } from 'node:fs/promises';
 import { colors } from './constants.js';
+import * as cmd from './commands.js';
 
 
 const args = process.argv[2]?.split('=');
@@ -19,22 +18,22 @@ let currentDir = homedir();
 console.log(`Welcome to the File Manager, ${colors.yellow}${userName}!${colors.reset}`);
 
 const readline = createInterface({ input: process.stdin, output: process.stdout });
-readline.on('close', onExit);
+readline.on('close', () => cmd.exit(userName));
 
 while (true) {
   const answer = await readline.question(`You are currently in ${colors.green}${currentDir}\n$>${colors.reset} `);
 
   switch (answer.trim().split(' ')[0]) {
     case '.exit':
-      onExit();
+      cmd.exit(userName);
       break;
 
     case 'up':
-      currentDir = onUp(currentDir);
+      currentDir = cmd.up(currentDir);
       break;
 
     case 'ls':
-      await onLs(currentDir);
+      await cmd.ls(currentDir);
       break;
 
     default:
@@ -42,45 +41,3 @@ while (true) {
       break;
   }
 }
-
-function onExit() {
-  console.log(`\nThank you for using File Manager, ${colors.yellow}${userName}${colors.reset}, goodbye!`);
-  process.exit();
-};
-
-function onUp(dir) {
-  if (dir.length === 2 && dir[1] === ':') return dir;
-  if (dir === '/') return dir;
-
-  return dir.split(sep).slice(0, -1).join(sep);
-}
-
-async function onLs(dir) {
-  try {
-    const dirContent = await readdir(dir);
-
-    const dirContentPromises = dirContent.map(async (item) => {
-      const path = join(dir, item);
-
-      try {
-        const itemStats = await stat(path);
-        return {
-          Name: item,
-          Type: itemStats.isDirectory() ? 'directory' : 'file'
-        };
-      } catch (error) {
-        return {
-          Name: item,
-          Type: 'file'
-        };
-      }
-    });
-
-    const result = await Promise.all(dirContentPromises);
-    console.table(result);
-
-  } catch (error) {
-    console.log(colors.red, 'Operation failed', colors.reset);
-  }
-}
-
